fix(cli): some template may not have specify files

+++ packages/cli/src/functional/AgenticaStarter.ts
@@ -178,7 +178,16 @@ export namespace AgenticaStarter {
     console.log("✅ Template downloaded");
 
     // REMOVE .GIT DIRECTORY
-    await fs.rm(path.join(directory, ".git"), { recursive: true });
-    await fs.rm(path.join(directory, ".github/dependabot.yml"));
+    // Some templates may not have .github/dependabot.yml
+
+    if (
+      await fs
+        .access(path.join(directory, ".github/dependabot.yml"))
+
+        .then(() => true)
+        .catch(() => false)
+    ) {
+      await fs.rm(path.join(directory, ".github/dependabot.yml"));
+    }
   };
 }
